--- 01-node-tutorial/answers/01-intro.js
@@ -1,9 +1,0 @@
-const amount = 9
-
-if (amount < 10) {
-  console.log('small number')
-} else {
-  console.log('large number')
-}
-
-console.log(`hey it's my first node app!!!`)
