--- 01-node-tutorial/answers/02-globals.js
@@ -1,5 +1,0 @@
-console.log(__dirname)
-setInterval(() => {
-  console.log('hello world')
-}, 1000)
-
