--- 01-node-tutorial/answers/05-utils.js
@@ -1,5 +1,0 @@
-const sayHi = (name) => {
-    console.log(`Hello there ${name}`)
-  }
-  // export default
-  module.exports = sayHi
