--- 01-node-tutorial/answers/06-alternative-flavors.js
@@ -1,6 +1,0 @@
-module.exports.items = ['item1', 'item2']
-const person = {
-  name: 'bob',
-}
-
-module.exports.singlePerson = person
