--- 01-node-tutorial/answers/08-os-module.js
@@ -1,16 +1,0 @@
-const os = require('os')
-
-// info about current user
-const user = os.userInfo()
-console.log(user)
-
-// method returns the system uptime in seconds
-console.log(`The System Uptime is ${os.uptime()} seconds`)
-
-const currentOS = {
-  name: os.type(),
-  release: os.release(),
-  totalMem: os.totalmem(),
-  freeMem: os.freemem(),
-}
-console.log(currentOS)
