--- 01-node-tutorial/answers/09-path-module.js
@@ -1,12 +1,0 @@
-const path = require('path')
-
-console.log(path.sep)
-
-const filePath = path.join('/content/', 'subfolder', 'test.txt')
-console.log(filePath)
-
-const base = path.basename(filePath)
-console.log(base)
-
-const absolute = path.resolve(__dirname, 'content', 'subfolder', 'test.txt')
-console.log(absolute)
