+++ 01-node-tutorial/answers/content/01-intro.js
@@ -1,0 +1,9 @@
+const amount = 9;
+
+if (amount < 10){
+    console.log('small number');
+} else{
+    console.log('large number');
+}
+
+console.log(`hey its my first node app!`)
