+++ 01-node-tutorial/answers/content/02-globals.js
@@ -1,0 +1,5 @@
+console.log(__dirname)
+
+setInterval(() => {
+    console.log('hello world')
+}, 1000);
