+++ 01-node-tutorial/answers/content/03-modules.js
@@ -1,11 +1,16 @@
+//modules
+//where module is located
+//importing modules
 const names = require('./04-names')
-
 const sayHi = require('./05-utils')
 
 const data = require('./06-alternative-flavor')
 
 require('./07-mind-grenade')
 
+
+
 sayHi('susan')
 sayHi(names.john)
 sayHi(names.peter)
+
