+++ 01-node-tutorial/answers/content/05-utils.js
@@ -1,0 +1,5 @@
+const sayHi = (name) => {
+    console.log(`Hello there ${name}`);
+}
+//exporting function export default
+module.exports = sayHi;
