+++ 01-node-tutorial/answers/content/06-alternative-flavor.js
@@ -1,0 +1,10 @@
+
+module.exports.items = ['item1', 'item2'];
+
+const person = {
+    name: 'bob',
+
+}
+
+//setting up property 
+//module.exports.singlePerson = person
