+++ 01-node-tutorial/answers/content/07-mind-grenade.js
@@ -1,0 +1,9 @@
+const num1 = 5;
+
+const num2 = 10;
+
+function addValues() {
+    console.log(`the sum is ${num1 + num2}`)
+}
+
+addValues()
