+++ 01-node-tutorial/answers/content/08-os-module.js
@@ -1,0 +1,25 @@
+//os module
+//built in
+const { openSync } = require('fs')
+const os = require('os')
+
+//info about current user
+
+const user = os.userInfo()
+
+console.log(user)
+
+
+//method returns the system uptime in seconds
+
+console.log(`The system uptime is ${os.uptime()} seconds`)
+
+const currentOS = {
+    name: os.type(),
+    release:os.release(),
+    totalMem: os.totalmem(),
+    freeMem: os/os.freemem(),
+    
+}
+
+console.log(currentOS)
