+++ 01-node-tutorial/answers/content/11-fs-async.js
@@ -1,0 +1,28 @@
+const {readFile, writeFile} = require('fs')
+
+console.log('start')
+//if no utf coding, get buffer
+readFile('./content/first.txt', 'utf8', (err, result)=>{
+    if(err){
+        console.log(err)
+        return;
+    }
+    const first = result;
+    readFile('./content/second.txt', 'utf8', (err, result) =>{
+        if(err){
+            console.log(err)
+            return;
+        }
+        const second = result
+        writeFile('./content/result-async.txt',
+        `here is the result: ${first}, ${second}`,
+        (err,result)=>{
+            if(err) {
+                console.log(err)
+                return;
+            }
+            console.log('done with this task')
+        })
+    })
+}) 
+console.log('starting next task')
