+++ 01-node-tutorial/answers/content/12-http.js
@@ -1,0 +1,23 @@
+//http
+
+const http = require('http');
+
+const server = http.createServer((req, res)=>{
+    if(req.url === '/'){
+        res.end('Welcome to our home page')
+    }
+    // res.write()
+    // res.end()
+    if(req.url === '/about'){
+        res.end('Here is our short history')
+    }
+    res.end(`
+    <h1>ooops!</h1>
+    <p>We can't seem to find the page you are looking for</p>
+    <a href="/">back home</a>
+    `)
+})
+
+server.listen(5000)
+//no exit
+
