+++ 01-node-tutorial/answers/content/result-mind-grenade.txt
@@ -1,0 +1,1 @@
+THIS IS AWESOME : Hello this is first text file Hello this is second text fileTHIS IS AWESOME : Hello this is first text file Hello this is second text fileTHIS IS AWESOME : Hello this is first text file Hello this is second text file
