+++ 01-node-tutorial/answers/content/result-sync.txt
@@ -1,0 +1,1 @@
+Here is the result : Hello this is first text file, Hello this is second text fileHere is the result : Hello this is first text file, Hello this is second text fileHere is the result : Hello this is first text file, Hello this is second text file
